+++ K8sUtils/public/Get-PodEvent.ps1
@@ -12,21 +12,31 @@ If set, don't return normal events
 K8s namespace to use, defaults to default
 
 .EXAMPLE
-An example
+Get-PodEvent -PodName mypod
+
+Get all events for pod mypod
+
+.EXAMPLE
+Get-PodEvent -PodName mypod -NoNormal -Namespace test
+
+Get all non-normal events for pod mypod in namespace test
 
 .OUTPUTS
-One or more event objects for the pod
+One or more event objects for the pod, $null if error
 #>
-function Get-PodEvent(
-    [Parameter(Mandatory = $true)]
-    [string] $PodName,
-    [switch] $NoNormal,
-    [string] $Namespace = "default"
+function Get-PodEvent {
+    param (
+        [CmdletBinding()]
+        [Parameter(Mandatory = $true)]
+        [string] $PodName,
+        [switch] $NoNormal,
+        [string] $Namespace = "default"
     )
-{
+    Write-Verbose "kubectl get events --namespace $Namespace --field-selector `"involvedObject.name=$PodName`" -o json"
+
     $events = kubectl get events --namespace $Namespace --field-selector "involvedObject.name=$PodName" -o json | ConvertFrom-Json
 
-    if (!$events) {
+    if ($LASTEXITCODE -ne 0 -or $null -eq $events) {
         return $null
     }
     if ($NoNormal) {