+++ K8sUtils/public/Get-PodStatus.ps1
@@ -3,7 +3,7 @@
 Get the status of the pods for a release
 
 .PARAMETER Selector
-K8s select for finding the pods
+K8s selector for finding the pods
 
 .PARAMETER ReplicaCount
 Number of pods to wait for that match the selector
@@ -20,14 +20,21 @@ K8s namespace to use, defaults to default
 .PARAMETER OutputFile
 File to write output to in addition to the console
 
-.PARAMETER IsJob
-Is this pod a job's pod
+.PARAMETER PodType
+If Pod, all containers must be running, otherwise for jobs all must be terminated ok
 
 .EXAMPLE
-An example
+$hookStatus = Get-PodStatus -Selector "job-name=$PreHookJobName" `
+                                            -Namespace $Namespace `
+                                            -OutputFile $tempFile `
+                                            -TimeoutSec 1 `
+                                            -PollIntervalSec $PollIntervalSec `
+                                            -PodType PreInstallJob
+
+Get the status of a pre-install job pod
 
 .OUTPUTS
-$True if all pods are running, $False if not
+Array of PodStatus object
 #>
 function Get-PodStatus {
 [CmdletBinding()]
@@ -41,7 +48,8 @@ param(
     [int] $TimeoutSec = 600,
     [string] $Namespace = "default",
     [string] $OutputFile,
-    [switch] $IsJob
+    [ValidateSet("Pod", "PreInstallJob", "Job")]
+    [string] $PodType = "Pod"
 
 )
 $ErrorActionPreference = 'Stop'
@@ -52,16 +60,22 @@ $runningPods = @{}
 $podStatuses = @{}
 $createdTempFile = !$OutputFile
 
-if ($IsJob) {
-    $phases = ,"Succeeded"
+if ($PodType -eq "PreInstallJob" ) {
+    $IsJob = $true
+    $okPhase = "Succeeded"
     $prefix = "preHook job pod"
+} elseif ($PodType -eq "Job") {
+    $IsJob = $true
+    $okPhase = "Succeeded"
+    $prefix = "job pod"
 } else {
-    $phases = ,"Running"
+    $IsJob = $false
+    $okPhase = ,"Running"
     $prefix = "pod"
 }
 
 # are all containers in the pod ready?
-function podReady($containerStatuses) {
+function allContainersReady($containerStatuses) {
 
     $readyContainers = @()
     Write-Verbose "ContainerStatuses: $($containerStatuses | ConvertTo-Json -Depth 10 -EnumsAsStrings)"
@@ -73,7 +87,7 @@ function podReady($containerStatuses) {
         $readyContainers += $containerStatuses | Where-Object ready -eq $true
     }
     $podIsReady = $readyContainers.Count -eq $containerStatuses.Count
-    Write-Verbose "Checking containerStatuses for $($IsJob ? 'job' : 'NON job'). PodReady = $podIsReady"
+    Write-Verbose "Checking containerStatuses for $($IsJob ? 'job' : 'NON job'). PodIsReady = $podIsReady"
     return $podIsReady
 }
 
@@ -87,6 +101,8 @@ $logSeconds = "600s"
 $extraSeconds = 1 # extra seconds to add to logSeconds to avoid missing something
 $lastEventTime = (Get-Date).AddMinutes(-5)
 $timedOut = $false
+
+Write-Status "Checking status of pods that match selector $Selector"
 while ($runningCount -lt $ReplicaCount -and !$timedOut)
 {
     $timedOut = (Get-Date) -gt $timeoutEnd
@@ -103,6 +119,11 @@ while ($runningCount -lt $ReplicaCount -and !$timedOut)
     $i = 0
     foreach ($pod in $pods) {
 
+        $i += 1
+        if ($runningPods[$pod.metadata.name]) {
+            continue # this pod is already completed
+        }
+
         if (!$podStatuses[$pod.metadata.name]) {
             $podStatuses[$pod.metadata.name] = [PodStatus]::new($pod.metadata.name)
         }
@@ -110,18 +131,13 @@ while ($runningCount -lt $ReplicaCount -and !$timedOut)
         $HasInit = [bool](Get-Member -InputObject $pod.spec -Name initContainers -ErrorAction SilentlyContinue)
         Write-Verbose "Pod $($pod.metadata.name) has init container: $HasInit."
 
-        if ($timedOut) {
-            Write-PodEvent -Prefix $prefix -PodName $pod.metadata.name -Namespace $Namespace -LogLevel ok -FilterStartupWarnings
-            Write-PodLog -Prefix $prefix -PodName $pod.metadata.name -Namespace $Namespace -LogLevel ok -HasInit:$HasInit
+        if (!(Get-Member -InputObject $pod.status -Name containerStatuses)) {
+            Write-Warning "Pod $($pod.metadata.name) has no status.containerStatuses. Pod:"
+            Write-Warning ($pod | ConvertTo-Json -Depth 10)
             continue
         }
 
-        $i += 1
-        if ($runningPods[$pod.metadata.name]) {
-            continue
-        }
-
-        $containers = $pod.status.containerStatuses.Count
+        $containers = @($pod.status.containerStatuses).Count
         $readyContainers = @($pod.status.containerStatuses | Where-Object ready -eq $true).Count
         Write-Verbose "ReplicaCount: $ReplicaCount RunningCount: $RunningCount PodCount: $($pods.Count)"
 
@@ -133,16 +149,17 @@ while ($runningCount -lt $ReplicaCount -and !$timedOut)
             $pod | ConvertTo-Json -Depth 10 | Out-File (Join-Path ([System.IO.Path]::GetTempPath()) "pod.json")
         }
 
-        Write-Verbose "Phases are $($phases -join ','). Phase is $($pod.status.phase)"
-        if ($pod.status.phase -in $phases) {
-            # "  $prefix $($pod.metadata.name) status is $($pod.status.phase)" | Tee-Object $OutputFile -Append | Write-MyHost
-            if (!$runningPods[$pod.metadata.name] -and
-                (podReady $pod.status.containerStatuses)) {
+        Write-Verbose "Ok phase is $okPhase. Pod's phase is $($pod.status.phase)"
+        if ($pod.status.phase -eq $okPhase) {
+            Write-Verbose "  $prefix $($pod.metadata.name) status is $($pod.status.phase)"
+            if (allContainersReady $pod.status.containerStatuses) {
 
                 $status = $IsJob ? [Status]::Completed : [Status]::Running
-                "  $prefix $($pod.metadata.name) has all containers ready or completed. Status is $status`n" | Tee-Object $OutputFile -Append | Write-MyHost
+                Write-Plain "  $prefix $($pod.metadata.name) has all containers ready or completed. Status is $status"
+
                 $runningCount += 1
                 $runningPods[$pod.metadata.name] = $true
+
                 $podStatuses[$pod.metadata.name].Status = $status
                 $podStatuses[$pod.metadata.name].ContainerStatuses = @($pod.status.containerStatuses | ForEach-Object { [ContainerStatus]::new($_.name, $status) })
                 if ($HasInit) {
@@ -155,17 +172,22 @@ while ($runningCount -lt $ReplicaCount -and !$timedOut)
                 Write-PodLog -Prefix $prefix -PodName $pod.metadata.name -Namespace $Namespace -LogLevel ok -HasInit:$HasInit
                 continue
             } else {
-                Write-Verbose "Pod $($pod.metadata.name) is ready, but pod containerStatuses are: $($pod.status.containerStatuses | out-string)"
+                Write-Verbose "Pod $($pod.metadata.name) is ready (phase = $okPhase), but pod containerStatuses are: $($pod.status.containerStatuses | out-string)"
             }
+        }
+
+        if ($timedOut) {
+            Write-PodEvent -Prefix $prefix -PodName $pod.metadata.name -Namespace $Namespace -LogLevel ok -FilterStartupWarnings
+            Write-PodLog -Prefix $prefix -PodName $pod.metadata.name -Namespace $Namespace -LogLevel ok -HasInit:$HasInit
+            break
         }
 
         # check for any errors since not ready yet
         $lastEventTime = Get-Date
-        Write-Verbose "kubectl get events --namespace $Namespace --field-selector `"involvedObject.name=$($pod.metadata.name)`" -o json"
-        $events = kubectl get events --namespace $Namespace --field-selector "involvedObject.name=$($pod.metadata.name)" -o json | ConvertFrom-Json
+        $events = Get-PodEvent -Namespace $Namespace -PodName $pod.metadata.name
         if ($events) {
-            $errors = @($events.items | Where-Object { $_.type -ne "Normal" -and $_.message -notlike "Startup probe failed:*"})
-            Write-Verbose "Got $($events.items.count) events for pod $($pod.metadata.name) "
+            $errors = @($events | Where-Object { $_.type -ne "Normal" -and $_.message -notlike "Startup probe failed:*" -and $_.reason -ne "FailedScheduling"})
+            Write-Verbose "Got $($events.count) events for pod $($pod.metadata.name) "
             Write-Verbose "Got $($errors.count) errors"
             if ($errors -or $pod.status.phase -eq "Failed" ) {
                 Write-Status "Pod $($pod.metadata.name) has $($errors.count) errors" -LogLevel Error
@@ -175,7 +197,13 @@ while ($runningCount -lt $ReplicaCount -and !$timedOut)
                 Write-PodLog -Prefix $prefix -PodName $pod.metadata.name -Namespace $Namespace -LogLevel Error -HasInit:$HasInit
 
                 # get latest pod status since sometimes get containerCreating status here
-                $pod = kubectl get pod --namespace $Namespace $pod.metadata.name -o json | ConvertFrom-Json
+                $name = $pod.metadata.name
+                $podJson = kubectl get pod --namespace $Namespace $name -o json
+                $pod = $podJson | ConvertFrom-Json
+                if (!$pod -or !(Get-Member -InputObject $pod -Name metadata)) {
+                    Write-Warning "Unexpected response from kubectl get pod --namespace $Namespace $name is: '$podJson'"
+                    throw "Unexpected response from kubectl get pod --namespace $Namespace $name"
+                }
 
                 $podStatuses[$pod.metadata.name].ContainerStatuses = @($pod.status.containerStatuses | ForEach-Object {
                         Write-Verbose "Pod status: $($_ | ConvertTo-Json -Depth 10)"
@@ -187,7 +215,7 @@ while ($runningCount -lt $ReplicaCount -and !$timedOut)
                 Write-Verbose "Get-PodStatus returning $($podStatuses[$pod.metadata.name] | ConvertTo-Json -Depth 10 -EnumsAsStrings)"
                 return $podStatuses.Values
             } else {
-                # "  No errors found for pod $($pod.metadata.name) yet`n" | Tee-Object $OutputFile -Append | Write-MyHost
+                Write-Verbose "No errors found for pod $($pod.metadata.name) yet"
 
                 # write intermediate events and logs for this pod
                 if ($VerbosePreference -eq 'Continue') {
@@ -196,15 +224,17 @@ while ($runningCount -lt $ReplicaCount -and !$timedOut)
                 }
             }
        }
+       # TODO we've seen case where pod.status.containerStatuses.state.waiting has
+       #   message: secret "eventhub-disabled-bootstrap-servers" not found
+       #   reason: CreateContainerConfigErrorreason: ImagePullBackOff
+       # but nothing in events. Local testing always has events.
     } # end foreach pod
 
     if ($runningCount -ge $ReplicaCount) {
         Write-Status "All ${prefix}s ($runningCount/$ReplicaCount) that matched selector $Selector are running`n" -Length 0 -Char '-' -LogLevel normal
         break
     }
-    if ($timedOut) {
-        break
-    }
+
     Write-Verbose "Sleeping $PollIntervalSec second$($PollIntervalSec -eq 1 ? '': 's'). Running Count = $runningCount ReplicaCount = $ReplicaCount"
     Start-Sleep -Seconds $PollIntervalSec
     $logSeconds = "$($PollIntervalSec + $extraSeconds)s"
@@ -213,7 +243,10 @@ while ($runningCount -lt $ReplicaCount -and !$timedOut)
 $ok = [bool]($runningCount -ge $ReplicaCount)
 if (!$ok) {
     Write-Verbose "Times: $(Get-Date) -lt $($timeoutEnd) Values count: $($podStatuses.Values.Count)"
-    Write-Status "Timed or errored out waiting $([int](((Get-Date) - $start).TotalSeconds))s for pods that matched selector $Selector RunningCount: $runningCount ReplicaCount: $ReplicaCount $ok" `
+    Write-Status "Error getting status after $([int](((Get-Date) - $start).TotalSeconds))s for pods that matched selector $Selector" `
+                -Length 0 `
+                -LogLevel Error
+    Write-Status "    RunningCount: $runningCount ReplicaCount: $ReplicaCount Ok: $ok TimedOut $timedOut" `
                 -Length 0 `
                 -LogLevel Error
     if ($podStatuses.Count -eq 0) {